Refactor trusted app registry: move AllowedOriginUrls and AppIdentifierToUrl definitions to trustedAppRegistry; update allowed origins to use dynamic URLs based on app identifiers.
