+++ src/TrabyterStaking_frontend/src/lib/javascript/Abstractions/constants/globalConstants.ts
@@ -1,20 +1,1 @@
-import {AppIdentifier} from '$lib/shared/common/abstractions/types/commonTypes';
-
-const InProduction: boolean = false;
-
 export const isInIframe = window.self !== window.top;
-
-export const TrabyterWebsiteUrl: string = InProduction
-    ? 'https://c42x7-waaaa-aaaap-qp3ba-cai.icp0.io/'
-    : 'http://u6s2n-gx777-77774-qaaba-cai.localhost:4943/';
-
-export const TrabyterStakingAppUrl: string = InProduction
-    ? 'https://2mjwp-daaaa-aaaak-qimya-cai.icp0.io'
-    : 'http://ucwa4-rx777-77774-qaada-cai.localhost:4943/';
-
-export const AllowedOriginUrls: string[] = [TrabyterWebsiteUrl, TrabyterStakingAppUrl];
-
-export const AppIdentifierToUrl: Partial<Record<AppIdentifier, string>> = {
-    [AppIdentifier.MainWebsite]: TrabyterWebsiteUrl,
-    [AppIdentifier.TrabyterStaking]: TrabyterStakingAppUrl,
-};
